requirements.txt
alembic
asyncpg
psycopg2
fastapi
uvicorn[standard]
httpx
sqlalchemy
pydantic
pydantic-settings
python-jose[cryptography]
python-multipart
argon2-cffi
pytest
pytest-cov
pytest_asyncio
redis